Add utils.plot_normalized_parameter and utils._parameter_normalized_subplot
-utils._parameter_normalized_subplot:
Takes a list of components and a parameter name, sums up all the
parameter values, and then plots each parameter normalized by this sum.
Useful for comparing the area in fine structure when the intensity
changes over for example a line scan.
-utils.plot_normalized_parameter:
Basically a "wrapper" for utils._parameter_normalized_subplot